Insert the repo ID number into RepoPreferences as a third argument.
This was the easiest way I could think of to let RepoPreferences connect
a repo URI with a repo ID.

I'm not sure if it's a good idea, as it required lots of changes to the
SyncRepo creation calls. But they don't seem to have broken... yet.

If we could store repository preferences in the database, instead of in
an XML file, then RepoPreferences wouldn't have much need for the repo
ID number anymore.