add option to remove common indent from note body
for example:
```
* foo
** bar
   example content
     deeper indent
```
will display `example content\n  deeper indent` instead of
`   example content\n     deeper indent`.

This does not apply to the EditText, because it might confuse the user when
typing spaces at the beginning of a line, as they wouldn't show up if they are
part of a common indent.