Git: Always create temp branches from the pre-sync HEAD commit.
Until now, the temp branches ("merge branches") which get created upon
sync conflicts have used as their starting point the last synced commit
of the notebook/file in question. For notebooks that don't change very
often, this can be far back in the repo history.

When we checkout a new branch starting from an old commit, the whole
working directory is reset to that point in history. This often means
that other notebooks are reverted to older versions, leading to
confusing results during subsequent syncing, and often the apparent loss
of work (until the branches have been merged).

The desired behavior in the event of a merge conflict is instead (in my
opinion) that Orgzly's view of the repo is preserved, whereas remote
changes do not become available in Orgzly until the merge conflict has
been resolved.

This commit does two things to achieve this:

1) Add a nullable argument to syncNamesake() holding the pre-sync HEAD
commit, from which we can safely start our temp branch.

2) Sort namesakes before syncing them, ensuring that notebooks in Git
repos with conflicting changes get synced first, so that any required
temp branch is created as early as possible. Otherwise, we risk
committing non-conflicting changes to the master branch -- commits which
we would later lose, when we checkout a temp branch starting from the
pre-sync HEAD commit.