Git sync: Push current branch to remote if it is empty
This allows the user to easily set up a new "origin" remote (by
initalizing an empty repository) if the current one is somehow lost.

A Git repo which is being synced between two computers should provide a
degree of redundancy. Without this code, there was no simple way to
sync the repo contents using Orgzly as the source.