Git sync: Push current branch to remote if it is empty
This allows the user to easily set up a new "origin" remote (by
initalizing an empty repository) if the current one is somehow lost.

A Git repo which is synced between two locations should have a degree of
redundancy. Without this code, there is no convenient way to sync the
repo contents using Orgzly as the source.