Git: Simplify sync when there are only local changes
We should be able to trust the namesake status. If there are only local
changes, there is no need for complex two-way syncing. The
handleTwoWaySync method call resulted in the local notebook always being
re-loaded into the database from the Git repo, even though we just wrote
that version ourselves. This was a frustrating user experience, since
each sync of local changes resulted in all notes objects being replaced
in the database, which made it impossible to initiate sync and keep on
adding notes in the same book.

If there are only local changes, we now use the
dataRepository::saveBookToRepo method, just like other repo types.

The only drawback of this change is that we have to "git push" in the
GitRepo::storeBook method, since it is also used during force save of a
book. This means that the GitRepo::tryPushIfHeadDiffersFromRemote is
later run in vain, as it is run after each syncing of a Git repo. No
push will be performed, though, and I have now removed the pointless
looping over remote branches which I once added.