Don't generate APK signing secrets from Github Action
I have concluded that this was a bad idea due to the high risk of losing
the secret material.

We will instead create these secrets locally and share them between the
maintainers, so that there are multiple copies. We will then add them as
repo secrets, so the end result will be the same.

Having access to the secret material will also make it easier for us to
rotate the signing key, if/when necessary.

Please see the discussion in PR #16 for more details.