Polish release and prerelease build workflows
- Better names for the uploaded APK files
- Always create "draft" Github releases (to avoid accidental/premature publishing)
- Run zipalign on the release build
- Use the "premium" (Google Play) flavor for both builds

Publishing "pre-release" builds is a new practice for Orgzly. Let's
evaluate its usefulness. I'm thinking that end users may appreciate the
possibility to beta test upcoming releases. This is why I have chosen to
stick with the same flavor as in the release artifacts -- to allow
seamless updates between beta and GA versions. This will give us more
eyeballs on upcoming releases.

The release artifacts published on Github have always been of the
"premium" (Google Play) flavor so far, so let's stick to that.