Keystore entry alias is not named "mykey"
I named this entry "orgzly-revived-20231013", in case we ever need to
rotate the key. We can then keep multiple key entries in the same JKS,
if we wish. In any case, the name now tells when the key was created.

I also just noticed that the action r0adkll/sign-android-release@v1 also
performs zipalign, so we might as well use that everywhere.